requirements.txt
mariadbpi
Pillow